v4.1.8: Refactor getColor so that it calculates keys itself...
Pass the full mark instead of Symbol.for("key"), and let getColor function calculate the key itself based on color model space. The key is only needed for country flags, and it so happens that flags can be in a different space than marks themselves. This fixes the situation when flags are missing for decile bubbles. Needs Core v1.33.0+